+++ tomcat/src/lxx/model/attributes/attribute_extractors/my/MyTravelTimeVE.java
@@ -13,8 +13,8 @@ import java.util.List;
 
 public class MyTravelTimeVE implements AttributeValueExtractor {
 
-    public int getAttributeValue(LXXRobot enemy, LXXRobot me, List<LXXBullet> myBullets) {
-        return LXXUtils.limit(0, (int) (me.getTime() - me.getLastStopTime()), 255);
+    public double getAttributeValue(LXXRobot enemy, LXXRobot me, List<LXXBullet> myBullets) {
+        return LXXUtils.limit(0D, (me.getTime() - me.getLastStopTime()), 255D);
     }
 
 }
